+++ front/lib/api/frames/move_source.ts
@@ -5,16 +5,11 @@ import type { GCSMountPoint } from "@app/lib/api/files/gcs_mount/files";
 import { emitGCSMountFileMovedAuditLog } from "@app/lib/api/files/gcs_mount/files";
 import { withFrameSourceAndPublishLock } from "@app/lib/api/frames/publication_storage";
 import {
-  MAX_FRAME_SOURCE_BYTES,
-  MAX_FRAME_SOURCE_FILE_COUNT,
-} from "@app/lib/api/frames/source_limits";
+  cleanupFrameSourceCopy,
+  copyFrameSourceAsNew,
+} from "@app/lib/api/frames/source_copy";
 import type { SandboxFunctionError } from "@app/lib/api/sandbox_functions/errors";
 import type { Authenticator } from "@app/lib/auth";
-import {
-  GCS_OBJECT_DOES_NOT_EXIST_GENERATION_MATCH,
-  getPrivateUploadBucket,
-} from "@app/lib/file_storage";
-import { isGCSPreconditionFailedError } from "@app/lib/file_storage/types";
 import { ConversationResource } from "@app/lib/resources/conversation_resource";
 import { FileResource } from "@app/lib/resources/file_resource";
 import type { FrameScopeTransitionStateError } from "@app/lib/resources/frame_sandbox_adapter";
@@ -23,7 +18,6 @@ import {
   FrameSandboxAdapter,
 } from "@app/lib/resources/frame_sandbox_adapter";
 import type { ScopeTransitionDestroyError } from "@app/lib/resources/sandbox_resource";
-import { concurrentExecutor } from "@app/lib/utils/async_utils";
 import logger from "@app/logger/logger";
 import { FRAME_MANIFEST_FILE } from "@app/types/api/frame_manifest";
 import type { ConversationWithoutContentType } from "@app/types/assistant/conversation";
@@ -33,10 +27,7 @@ import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import { assertNever } from "@app/types/shared/utils/assert_never";
 import { normalizeError } from "@app/types/shared/utils/error_utils";
-import type { File, FileMetadata } from "@google-cloud/storage";
 import { UniqueConstraintError } from "sequelize";
-
-const FRAME_SOURCE_MOVE_COPY_CONCURRENCY = 4;
 
 type FrameSourceOwner = {
   useCase: FileUseCase;
@@ -181,197 +172,6 @@ function assertFrameAtSource(
     );
   }
   return new Ok(undefined);
-}
-
-function metadataValue(metadata: FileMetadata, key: keyof FileMetadata) {
-  const value = metadata[key];
-  return value === undefined || value === null ? null : String(value);
-}
-
-function isSameGCSObject(source: File, destination: File): boolean {
-  const sourceMd5 = metadataValue(source.metadata, "md5Hash");
-  const destinationMd5 = metadataValue(destination.metadata, "md5Hash");
-  if (sourceMd5 && destinationMd5) {
-    return sourceMd5 === destinationMd5;
-  }
-
-  const sourceCrc32c = metadataValue(source.metadata, "crc32c");
-  const destinationCrc32c = metadataValue(destination.metadata, "crc32c");
-  return Boolean(
-    sourceCrc32c &&
-      destinationCrc32c &&
-      sourceCrc32c === destinationCrc32c &&
-      metadataValue(source.metadata, "size") ===
-        metadataValue(destination.metadata, "size")
-  );
-}
-
-type DestinationObject = { filePath: string; generation: string };
-type FrameSourceCopyFailure = {
-  destinationObjects: DestinationObject[];
-  error: FrameSourceMoveError;
-};
-
-async function cleanupDestinationObjects(
-  objects: DestinationObject[]
-): Promise<boolean> {
-  const bucket = getPrivateUploadBucket();
-  try {
-    await concurrentExecutor(
-      objects,
-      ({ filePath, generation }) =>
-        bucket.delete(filePath, {
-          ignoreNotFound: true,
-          ifGenerationMatch: generation,
-        }),
-      { concurrency: FRAME_SOURCE_MOVE_COPY_CONCURRENCY }
-    );
-    return true;
-  } catch (error) {
-    logger.error(
-      { error: normalizeError(error) },
-      "Failed to clean up a Frame move destination"
-    );
-    return false;
-  }
-}
-
-async function copyFrameSourceAsNew({
-  allowMatchingDestinationObjects,
-  destinationMountPrefix,
-  sourceMountPrefix,
-}: {
-  allowMatchingDestinationObjects: boolean;
-  destinationMountPrefix: string;
-  sourceMountPrefix: string;
-}): Promise<Result<DestinationObject[], FrameSourceCopyFailure>> {
-  const bucket = getPrivateUploadBucket();
-  const [sourceFiles, destinationFiles] = await Promise.all([
-    bucket.getFiles({
-      prefix: sourceMountPrefix,
-      maxResults: MAX_FRAME_SOURCE_FILE_COUNT + 1,
-    }),
-    bucket.getFiles({
-      prefix: destinationMountPrefix,
-      maxResults: MAX_FRAME_SOURCE_FILE_COUNT + 1,
-    }),
-  ]);
-  if (sourceFiles.length === 0) {
-    return new Err({
-      destinationObjects: [],
-      error: new FrameSourceMoveError(
-        "invalid_source",
-        "Frame source folder is empty or no longer exists."
-      ),
-    });
-  }
-  const totalSourceSize = sourceFiles.reduce(
-    (total, file) => total + Number(file.metadata.size ?? 0),
-    0
-  );
-  if (
-    sourceFiles.length > MAX_FRAME_SOURCE_FILE_COUNT ||
-    totalSourceSize > MAX_FRAME_SOURCE_BYTES
-  ) {
-    return new Err({
-      destinationObjects: [],
-      error: new FrameSourceMoveError(
-        "invalid_source",
-        "Frame source exceeds the move size limit."
-      ),
-    });
-  }
-  if (destinationFiles.length > MAX_FRAME_SOURCE_FILE_COUNT) {
-    return new Err({
-      destinationObjects: [],
-      error: new FrameSourceMoveError(
-        "conflict",
-        "Frame destination exceeds the move file count limit."
-      ),
-    });
-  }
-
-  const destinationByRelativePath = new Map(
-    destinationFiles.map((file) => [
-      file.name.slice(destinationMountPrefix.length),
-      file,
-    ])
-  );
-  const copyResults = await concurrentExecutor(
-    sourceFiles,
-    async (sourceFile) => {
-      try {
-        const relativePath = sourceFile.name.slice(sourceMountPrefix.length);
-        const destinationPath = `${destinationMountPrefix}${relativePath}`;
-        let destinationFile = destinationByRelativePath.get(relativePath);
-        const destinationWasListed = Boolean(destinationFile);
-        let copied = false;
-
-        if (!destinationFile) {
-          try {
-            await bucket.copyFile(sourceFile.name, destinationPath, undefined, {
-              destinationGenerationMatch:
-                GCS_OBJECT_DOES_NOT_EXIST_GENERATION_MATCH,
-            });
-            copied = true;
-          } catch (error) {
-            if (!isGCSPreconditionFailedError(error)) {
-              throw error;
-            }
-          }
-          destinationFile = bucket.file(destinationPath);
-          const [metadata] = await destinationFile.getMetadata();
-          destinationFile.metadata = metadata;
-        }
-
-        if (
-          (!copied && !isSameGCSObject(sourceFile, destinationFile)) ||
-          (!allowMatchingDestinationObjects && destinationWasListed)
-        ) {
-          throw new FrameSourceMoveError(
-            "conflict",
-            `Destination file already exists: ${relativePath}`
-          );
-        }
-
-        const generation = metadataValue(
-          destinationFile.metadata,
-          "generation"
-        );
-        if (!generation) {
-          throw new FrameSourceMoveError(
-            "internal",
-            `Destination generation is missing: ${relativePath}`
-          );
-        }
-        return new Ok<DestinationObject>({
-          filePath: destinationPath,
-          generation,
-        });
-      } catch (error) {
-        return new Err(
-          error instanceof FrameSourceMoveError
-            ? error
-            : new FrameSourceMoveError(
-                "internal",
-                `Failed to copy the Frame source: ${normalizeError(error).message}`
-              )
-        );
-      }
-    },
-    { concurrency: FRAME_SOURCE_MOVE_COPY_CONCURRENCY }
-  );
-  const destinationObjects = copyResults
-    .filter((result) => result.isOk())
-    .map((result) => result.value);
-  const failedCopy = copyResults.find((result) => result.isErr());
-  if (failedCopy?.isErr()) {
-    return new Err({
-      destinationObjects,
-      error: failedCopy.error,
-    });
-  }
-  return new Ok(destinationObjects);
 }
 
 /** Move a registered Frame folder while retaining its stable FileResource identity. */
@@ -593,11 +393,13 @@ export async function moveFrameV2Source(
     const copyResult = await copyFrameSourceAsNew({
       allowMatchingDestinationObjects: isRecovery,
       destinationMountPrefix,
+      makeError: (code, message) => new FrameSourceMoveError(code, message),
       sourceMountPrefix,
     });
     if (copyResult.isErr()) {
-      const cleaned = await cleanupDestinationObjects(
-        copyResult.error.destinationObjects
+      const cleaned = await cleanupFrameSourceCopy(
+        copyResult.error.destinationObjects,
+        { operation: "move" }
       );
       if (!cleaned) {
         return new Err(
@@ -676,7 +478,9 @@ export async function moveFrameV2Source(
           });
 
     if (updateResult.isErr()) {
-      const cleaned = await cleanupDestinationObjects(copyResult.value);
+      const cleaned = await cleanupFrameSourceCopy(copyResult.value, {
+        operation: "move",
+      });
       const restored = cleaned ? await restoreSourceReservation() : false;
       if (!restored) {
         logger.error(
